Re-create Awesome books app
