+++ awesome-books-with-ES6/modules/delete.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 import Storage from './Storage.js';
 import DummyData from './DummyData.js';
 import { bookContainer } from '../index.js';
